requirements.txt
tqdm
wandb
seaborn
scikit-learn==1.3.0
pytorch-lightning==1.8.6 
tabulate==0.9.0
hydra-core==1.3.1
opencv-python==4.6.0.66
timm==0.6.13
einops==0.3.0
transforms3d==0.4.1
piq==0.8.0
lpips==0.1.4
